pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["ncheymbamalu <nchey.learnings@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.14"
catboost = "1.2.5"
discordwebhook = "1.0.3"
fire = "0.6.0"
geopandas = "1.0.1"
hopsworks = "3.7.0"
ipykernel = "^6.29.2"
jupyter = "^1.0.0"
lightgbm = "4.4.0"
loguru = "^0.7.2"
nbformat = "5.10.4"
numpy = "1.25.2"
omegaconf = "2.3.0"
optuna = "3.6.1"
pandas = "2.0.3"
plotly = "5.15.0"
polars = "1.1.0"
pyarrow = "16.1.0"
pydeck = "0.9.1"
python-dotenv = "1.0.1"
requests = "2.31.0"
retry = "0.9.2"
scikit-learn = "1.2.2"
streamlit = "1.36.0"
tqdm = "4.66.4"
xgboost = "2.1.0"

[tool.poetry.dev-dependencies]
pip-tools = "6.13.0"
pytest = "7.4.4"
watchdog = "4.0.1"

[tool.poetry.group.dev.dependencies]
black = "24.4.2"
Flake8-pyproject = "1.2.3"
isort = "5.13.2"
mypy = "1.10.1"
pylint = "3.2.5"

[tool.black]
line-length = 100

[tool.flake8]
max-line-length = 100
import-order-style = ["pep8"]
application_import_names = ["src"]
exclude = [".git", "__pycache__"]
extend-ignore = ["E501", "W503", "W504", "E203", "I201", "I202"]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
indent = 4
lines_between_types = 1
known_first_party = ["src"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.10"
show_error_codes = true
ignore_missing_imports = true
install_types = true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true
warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true
verbosity = 0


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
